+++ assets/javascripts/kitten/karl/lazy/lazy-hook.js
@@ -5,7 +5,11 @@ const LazyHook = ({ children }) => {
   const lazyComponentRef = useRef(null)
   const isVisible = useLazyObserver(lazyComponentRef)
 
-  return <div ref={lazyComponentRef}>{isVisible && children}</div>
+  return (
+    <div ref={lazyComponentRef} style={{ background: '#eee' }}>
+      Here: {isVisible && children}
+    </div>
+  )
 }
 
 export default LazyHook